Summary: Updating Markdown, adding section titles, introducing `;`, adding .py and .html files
Hey - Adding a few things that might be helpful.

1) Markdown: Markdown can be super helpful in organizing the notebook -
I'm a fan of having a section at the top that lists the title of the notebook / analysis, has a brief description / overview,
and has links to resources. You can also link to points within the notebook, for easier navigation.

2) Adding section titles: Building off of adding a Markdown section in the beginning,
section titles help split up the notebook and improve organization.

3) Introducing `;`: Self-explanatory within the notebook

4) Adding `.py` and `.html` files: using the jupyter config explained in `projectRomeo/projectEnergy` to
automatically create a `.py` and `.html` version of the notebook for easier versioninng. Read more about
the config in `projectRomeo/projectEnergy/README.md`.